requirements.txt
flask
flask-cors
tensorflow
numpy
opencv-python
Pillow
scikit-learn
